+++ packages/blocks-core/CHANGELOG.md
@@ -34,6 +34,7 @@
 - `Heading` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `List` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `List Item` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
+- `Navigation Link` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 
 
 ### Automated Tests
@@ -69,6 +70,7 @@
 - Added E2E tests to check functionality `Heading` new inner block.
 - Added E2E tests to check functionality `List` new inner block.
 - Added E2E tests to check functionality `List Item` new inner block.
+- Added E2E tests to check functionality `Navigation Link` new inner block.
 
 ## 1.4.0 (2025-05-19)